Add Weight text in the README
This is not yet implemented, but this is how weight will be used in SRV
records. It might be worth while to extend this to A/AAAA replies
as well to get some load balancing going there too. However this needs
some local counters on how often some names have been given out. Not
hard, but this code isn't there yet.