-machines flag doesn't have expected behavior
Hey I think that this is what you were intending this flag to do. Currently it doesn't look like it does anything. I think that it was just a miss placed } and a little miss on setting the var because of that. Let me know if this needs more of an explanation. Thanks for this project im liking it so far.