Also compress when sending replies from cache.
When we return a message from cache we don't set the Compress to
true. In fact returning message from is always tricky because we don't
really know for sure if the requesting client is using the same buf
size as the original requesting client.

Anyway, not setting Compress to true is certainly a mistake here
as we do it for the normal (non-cache) replies.

See issue #129 and https://github.com/miekg/dns/issues/184