Use qname to synthesis target for SRV
When we generate the SRV Target when the Host is an IP we use the key
path turned into a Domain. This commit changes this into the Qname,
which is probable better.

Also the CNAME loop code is not correct, for now it is commented out,
as I need to sort this out.