Extend the testcase for issue #144
Add a testcase that highlights the issue. If commented out the original
change that would make this work as intended, i.e. right now the
target of the SRV is the key path name. This makes more sense as it
allows you to create loops. I'm also sure this is in use in the wild,
so this change would be backward incompatible...