Implement TargetFromQuery
A boolean flag on a msg which says:

When a SRV record with a "Host: IP-address" is added, we synthesize a srv.Target. Normally
we convert the Key where the record lives to a DNS name and use this as the srv.Target.
When TargetFromQuery is true we use the query name as the srv.Target.