Suppress duplicate CNAMEs
Fixes issue #168
Duplicate CNAMEs like:

    a.ipaddr.skydns.test.   3600    IN  A   172.16.1.1
    b.ipaddr.skydns.test.   3600    IN  A   172.16.1.2
    mx.skydns.test. 3600    IN  CNAME   a.ipaddr.skydns.test.
    mx.skydns.test. 3600    IN  CNAME   b.ipaddr.skydns.test.

are not allowed in the DNS and should be suppressed. Right now we just
allow one in the non-round-robin case the latter one (b.ipaddr.. in
this case). If round-robin is enabled we just flip a coin on wether to
include the "new" CNAME are not. The A/AAAA record that is not
referenced is still included in the answer, but this should not case any
harm, and should be ignored by the client.