skydns: add prometheus
Add prometheus suppport. Still need to litter the code with counter
increments and maybe add some tests.

I'm in favor of making prometheus the *only* metrics for SkyDNS, but
for now I'm leaving the Graphite support in, although I consider that
legacy. I also did not extend on the stats.Counter idea and I'm just
using the prometheus stuff directly.