Drop additinal section when msg > 64K
When a message is *really* too big, kill the entire additional section
and see if it will fit then.

Also fix a thing where we can't unmarshal json and still return NOERROR,
now we are going to return SERVFAUL.

And make the the overflow test really overflow by adding 2000 records
under one name. This takes a long time to run now.