Use the correct name when creating add. section records.
Bugfix: Use the srv.Target name for the name of the A and AAAA record
that are created in the additional section. Add test for this as well
'TestTargetStripAdditional' that tests for this when targetstrip is
in effect (but even with TargetStrip = 0 this was an issue).

Also factor out TargetStrip into a function.