Fix and cleanup caching.
Fix and cleanup caching. Cache differently for dnssec/non-dnssec and
UDP/TCP. Also add Fit function that trims a message to size (setting
TC when needed).

Drop the LRU caching a go with random eviction, eliminate all data races
and drop some of the useless cache metrics.

Add testing for the cache subpackage.

Cache stubforwarding, recursive and reverse requests.

Don't call EvictRandom under a lock This leads to deadlocks when the
cache is almost full. Drop the call to cache.Remove and just delete in
the map directly.